JSP_0924

+++ board22/src/main/java/article/service/DeleteArticleService.java
@@ -18,16 +18,30 @@ public class DeleteArticleService {
 		Connection conn = null;
 		
 		try {
+			// 커넥션 연결
 			conn = ConnectionProvider.getConnection();
+			
+			// 트랜잭션 시작
 			conn.setAutoCommit(false);
 			
+			// DAO에서 Article 객체를 얻어옴
 			Article article = articleDao.selectById(conn, delReq.getArticleNumber());
+			
+			// 해당 게시글이 존재하지 않으면 예외처리
 			if(article == null) {
 				throw new ArticleNotFoundException();
 			}
 			
+			// 해당 게시글의 글쓴이가 아니라면 예외처리
+			if(!canDelete(delReq.getUserId(), article)) {
+				throw new PermissionDeniedException();
+			}
+			
+			// delete 실행
 			articleDao.delete(conn, delReq.getArticleNumber());
 			contentDao.delete(conn, delReq.getArticleNumber());
+			
+			// 트랜잭션 종료
 			conn.commit();
 		} catch (SQLException e) {
 			JdbcUtil.rollback(conn);
@@ -38,5 +52,9 @@ public class DeleteArticleService {
 		} finally {
 			JdbcUtil.close(conn);
 		}
+	}
+	
+	private boolean canDelete(String modifyingUserId, Article article) {
+		return article.getWriter().getId().equals(modifyingUserId);
 	}
 }
